+++ descarteslabs/dynamic_compute/compute_map.py
@@ -140,6 +140,10 @@ def dot(*args, **kwargs):
     return None
 
 
+def from_image_ids(*args, **kwargs):
+    return None
+
+
 class ComputeMap(dict, ABC):
     """
     A wrapper class to support operations on grafts. Proxy objects should all be
@@ -214,6 +218,7 @@ class ComputeMap(dict, ABC):
                     ("functional", functional),
                     ("dot", dot),
                     ("filled", fill_mask),
+                    ("from_image_ids", from_image_ids),
                 ],
                 debug=True,
             )()